refactor(solver): split conditional phase helpers
